requirements.txt
boto3
Django
djangorestframework
django-crontab
django-environ
django-storages
django-filter
ipython